requirements.txt
instagrapi
pillow
python-docx
pypdf
